requirements.txt
Django==4.1.6
environs==9.5.0
django-debug-toolbar=3.8.1
folium==0.14.0
django-tinymce=3.5.0
Pillow=9.4.0
